@@ -5,30 +5,30 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using DualAuthCore.Models;
+using Microsoft.Extensions.Logging;
 
 namespace DualAuthCore.Controllers
 {
   public class HomeController : Controller
   {
+    private readonly ILogger<HomeController> _logger;
+
+    public HomeController(ILogger<HomeController> logger)
+    {
+      _logger = logger;
+    }
+
     public IActionResult Index()
     {
       return View();
     }
 
-    public IActionResult About()
+    public IActionResult Privacy()
     {
-      ViewData["Message"] = "Your application description page.";
-
       return View();
     }
 
-    public IActionResult Contact()
-    {
-      ViewData["Message"] = "Your contact page.";
-
-      return View();
-    }
-
+    [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
     public IActionResult Error()
     {
       return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
