@@ -1,6 +1,3 @@
-@using Microsoft.AspNetCore.Identity
-@using DualAuthCore
+﻿@using DualAuthCore
 @using DualAuthCore.Models
-@using DualAuthCore.Models.AccountViewModels
-@using DualAuthCore.Models.ManageViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
